Change linker script
This script is the one generated by the STM32CubeMX 6.1.1 for a basic
project, which is almost identical (even the comments) to the one
copyrighted by Attolic. Question is who copied it first.

This changes what is the highest address of the user mode stack, which
were generating weird effects using some HAL SPI drivers. The main one
was generating continuous clock and no data on MOSI/MISO pins.